pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-data-engineering"
version = "0.1.0"
description = "AI Data Engineering project with ETL, PostgreSQL, and AI Agent"
authors = [{name = "Rodrigo", email = "rodrigo@example.com"}]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "pandas>=2.1.4",
    "numpy>=1.24.3",
    "python-dotenv>=1.0.0",
    "psycopg2-binary>=2.9.9",
    "sqlalchemy>=2.0.23",
    "kaggle>=1.5.16",
    "kagglehub>=0.2.7",
    "openai>=1.3.7",
    "langchain>=0.0.350",
    "langchain-openai>=0.0.2",
    "mcp>=0.1.0",
    "streamlit>=1.28.2",
    "plotly>=5.17.0",
    "altair>=5.2.0",
    "requests>=2.31.0",
    "python-multipart>=0.0.6"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0"
]

[tool.black]
line-length = 88
target-version = ['py39']

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
